movie_recommender_be/requirements.txt
pandas
matplotlib
bs4
requests
tmdbv3api
python-dotenv
nltk
scikit-learn
fastapi
uvicorn